Pair programming with @travs.  Continuing breakdown of test.py: 1) DataTest 2) PintTest 3) QuantityTest 4) RDFLibTest
